Rename matrix_utils -> linalg
