+++ pyro/ops/linalg.py
